add user feedback

+++ ese2021-project-scaffolding/scaffolding/frontend/src/app/board/post/post.component.ts
@@ -6,6 +6,8 @@ import {environment} from "../../../environments/environment";
 import {PostCategory} from "../../models/postCategory.model";
 import {Router} from "@angular/router";
 import {ToastrService} from "ngx-toastr";
+import {MatDialog, MatDialogRef} from "@angular/material/dialog";
+import {ConfirmCancel} from "../../profile/orders/orders.component";
 
 @Component({
   selector: 'app-post',
@@ -59,7 +61,8 @@ export class PostComponent implements OnInit {
     public httpClient: HttpClient,
     public userService: UserService,
     public router: Router,
-    private toastr: ToastrService
+    private toastr: ToastrService,
+    public dialog: MatDialog
   ) {
     userService.loggedIn$.subscribe((res) => {this.loggedIn = res; this.whoCanVote(); this.whoCanEdit();});
     this.loggedIn = userService.getLoggedIn();
@@ -226,21 +229,28 @@ export class PostComponent implements OnInit {
     else{
     this.sendUpdate.emit(this.post);}
     this.editMode = false;
+    this.toastr.show("Post was updated");
   }
 
   deletePost(): void{
-    if (this.userService.isAdmin()) {
-      this.httpClient.delete(environment.endpointURL + "post/admin/" + this.post.postId  + "/" +this.post.creatorId)
-        .subscribe((()=>{
-          this.getNewPosts.emit();
-        }));
-    }
-    else {
-      this.httpClient.delete(environment.endpointURL + "post/user/" + this.post.postId  + "/" +this.userService.getUser().userId)
-      .subscribe((()=>{
-        this.getNewPosts.emit();
-      }));
-    }
+    const dialogRef = this.dialog.open(ConfirmDelete);
+    dialogRef.afterClosed().subscribe((cancel) => {
+      if(cancel){
+        if (this.userService.isAdmin()) {
+          this.httpClient.delete(environment.endpointURL + "post/admin/" + this.post.postId  + "/" +this.post.creatorId)
+            .subscribe((()=>{
+              this.getNewPosts.emit();
+            }));
+        }
+        else {
+          this.httpClient.delete(environment.endpointURL + "post/user/" + this.post.postId  + "/" +this.userService.getUser().userId)
+            .subscribe((()=>{
+              this.getNewPosts.emit();
+            }));
+        }
+        this.toastr.show("Post was deleted");
+      }
+    });
   }
 
   getCreator(){
@@ -267,4 +277,26 @@ export class PostComponent implements OnInit {
   validate(): boolean{
     return (this.post.title != '' &&( this.post.content!='' || (this.post.itemImage || this.selectedFile != null)));
   }
+}
+@Component({
+  selector: 'delete-post-confirm',
+  template: '<h2>Do you really want to delete this post?</h2>' +
+    '<button mat-flat-button color="warn" style="  margin: 5px; position: center;" (click)="delete()">Yes</button>' +
+    ' <button mat-flat-button color="accent" style="  margin: 5px; position: center;" (click)="dont()">No</button>',
+})
+
+export class ConfirmDelete {
+
+  constructor(
+    public dialogRef: MatDialogRef<ConfirmDelete>) { }
+
+
+  delete(): void {
+    this.dialogRef.close(true);
+  }
+
+  dont(): void{
+    this.dialogRef.close(false);
+  }
+
 }
